+++ CHANGELOG.md
@@ -7,6 +7,13 @@
   the application configuration (or passed to `pg:start_pool/2`). For example,
   a pool identified as `example` will have a process named `pg_pool_example`.
   It is therefore easier to guarantee name unicity.
+- Add `pg:pool_stats/1`, `pg:acquire/1`, `pg:release/2`, `pg:with_client/2`,
+  `pg:with_transaction/2`, which takes a pool identifier as argument instead
+  of a pool reference.
+- Add `pg:exec/2`, `pg:exec/3`, `pg:exec/4`, `pg:query/2`, `pg:query/3`,
+  `pg:query/4`. They simply call the equivalent functions from `pg_client`.
+## Bugs
+- Fix various invalid types.
 ## Misc
 - Log domains are now `[pg, client]` for clients and `[pg, pool]` for pools.
 - Query options are now the empty map by default, making